refactor: Correct dagger hilt error for AnalysisWorker
